optimizes madd functionality

+++ src/primitives.rs
@@ -18,5 +18,5 @@ pub type Degree = u32;
 pub type BasisLength = usize;
 
 // Matrix stuff
-pub type DenseRowCoefficient = i128;
+pub type DenseRowCoefficient = i64;
 pub type DenseRow = Vec<DenseRowCoefficient>;
